update storage data on bulk deleting folders

+++ views.py
@@ -177,12 +177,23 @@ class FolderViewSet(viewsets.ModelViewSet):
             arr = data["folder_id_arr"]
         except KeyError:
             return HttpResponse("folder ids not found.", status=status.HTTP_400_BAD_REQUEST)
-        try:
-            folders = Folder.objects.filter(pk__in=arr)
-            folders.delete()
-            return HttpResponse(status=status.HTTP_204_NO_CONTENT)
-        except:
-            return HttpResponse("error in deliting folders", status=status.HTTP_400_BAD_REQUEST)
+        # try:
+        folders = Folder.objects.filter(pk__in=arr)
+        if len(folders) == 0:
+            return HttpResponse("no folder ids given", status=status.HTTP_400_BAD_REQUEST)
+        total_folder_size = 0
+        for folder in folders:
+            total_folder_size = total_folder_size + folder.content_size
+        parent = folders[0].parent
+        while not parent == None:
+            updated_size = parent.content_size - total_folder_size
+            parent.content_size = updated_size
+            parent.save()
+            parent = parent.parent
+        folders.delete()
+        return HttpResponse(status=status.HTTP_204_NO_CONTENT)
+        # except:
+        #     return HttpResponse("error in deliting folders", status=status.HTTP_400_BAD_REQUEST)
 
     def destroy(self, request, *args, **kwargs):
         instance = self.get_object()
@@ -428,12 +439,13 @@ class AllSharedItems(APIView):
         }
         return Response(serializer)
 
+
 class AllStarredItems(APIView):
     """
     This view allows user to view all the starred items
     """
 
-    def get(self,request,*args, **kwargs):
+    def get(self, request, *args, **kwargs):
         filemanager_name = request.query_params.get('filemanager', None)
         try:
             filemanager = FileManager.objects.get(
@@ -442,14 +454,14 @@ class AllStarredItems(APIView):
             return Response("Filemanager instance with given name doesnot exists", status=status.HTTP_400_BAD_REQUEST)
         person = self.request.person
         files_starred = File.objects.filter(
-            folder__filemanager = filemanager).filter(folder__person = person).filter(starred = True
-        )
+            folder__filemanager=filemanager).filter(folder__person=person).filter(starred=True
+                                                                                  )
         files = FileSerializer(
             files_starred, many=True
         )
         folders_starred = Folder.objects.filter(
-            filemanager = filemanager).filter(person = person).filter(starred = True
-        )
+            filemanager=filemanager).filter(person=person).filter(starred=True
+                                                                  )
         folders = FolderSerializer(
             folders_starred, many=True
         )
@@ -460,6 +472,7 @@ class AllStarredItems(APIView):
             'filemanager': filemanager_name
         }
         return Response(serializer)
+
 
 class ItemSharedView(APIView):
     """
